Float types are currently spread out. this moves them to a single place to ensure we properly handle the public type interface in all three renderers
